refactor(core): move the spike-era proof and discovery modes out of the write engine
WriteEngine.cs opened with the step-4 build plan and a DefaultSourcePath
hardcoded to one machine's Steam install - a path that does not exist on the
dev machine - followed by console modes nothing on the tool surface reaches
(#453). A 2026-05 baseline review (finding N3) had already asked for the split.

Moved to housecarl-generator/WriteDiagnosticsProbe.cs, beside the probes they
resemble: npc-skills (the step-4 acceptance proof), write-api (Mutagen
override-surface discovery), poly-probe, substruct-probe - with the four helpers
nothing else used. Calls that were siblings inside WriteEngine are now qualified
calls into it; all are public or internal, and housecarl-core already declares
InternalsVisibleTo("housecarl-generator"), so nothing widened.

DefaultSourcePath is gone from housecarl-core. Its last user, condition-patch,
now requires --source like its siblings patch and show already do, instead of
failing with a path the caller never typed.

The class header is rewritten per DOC_HYGIENE section 8: what the engine is, the
contracts it keeps, the probes that prove them - and an explicit list of what did
NOT move (the patch/show/condition-patch dev harnesses, one of which drives
insert-at-index-guard, and the two coerce-* ci-all probes), so the residue is
named rather than left for the next reader to find.

Verified: ci-all 132/132; all four relocated modes run identically against a real
Skyrim.esm (npc-skills still reports PASS); condition-patch checked in both
directions - refused by name without --source, writes and reads back the
re-targeted condition with it.

Fixes #453
